페사드 계층 추가 및 API Gateway 의존성 수정 - `AuthController`, `UserController` 추가: 인증 및 사용자 관리 엔드포인트 구현. - `application.yml` 추가: 서비스 및 포트 설정 정의. - Gradle 설정 업데이트: `facade` 모듈에 Spring Boot 추가, API Gateway에서 `facade` 의존성 제거.
