+++ apps/facade/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'base-spring.feign'
+    id 'org.springframework.boot'
 }
 
 dependencies {
