+++ apps/facade/src/main/java/com/petfeed/facade/controller/AuthController.java
@@ -1,0 +1,24 @@
+package com.petfeed.facade.controller;
+
+import com.petfeed.domain.dto.request.LoginRequest;
+import com.petfeed.domain.dto.response.LoginResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/v1/auth")
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final AuthFacade authFacade;
+
+    @PostMapping("/login")
+    public ResponseEntity<LoginResponse> login(@Valid @RequestBody LoginRequest request) {
+        return ResponseEntity.ok(authFacade.login(request));
+    }
+}
